Rename ServerTelemetry to EstsTelemetry
