+++ common/src/main/java/com/microsoft/identity/common/internal/providers/oauth2/OAuth2Strategy.java
@@ -35,7 +35,7 @@ import com.microsoft.identity.common.internal.net.HttpResponse;
 import com.microsoft.identity.common.internal.net.ObjectMapper;
 import com.microsoft.identity.common.internal.platform.Device;
 import com.microsoft.identity.common.internal.providers.microsoft.MicrosoftTokenRequest;
-import com.microsoft.identity.common.internal.servertelemetry.ServerTelemetry;
+import com.microsoft.identity.common.internal.eststelemetry.EstsTelemetry;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -144,7 +144,7 @@ public abstract class OAuth2Strategy
             );
         }
         headers.putAll(Device.getPlatformIdParameters());
-        headers.putAll(ServerTelemetry.getTelemetryHeaders());
+        headers.putAll(EstsTelemetry.getTelemetryHeaders());
 
         return HttpRequest.sendPost(
                 new URL(mTokenEndpoint),
